requirements.txt
python-telegram-bot==13.15
solana==0.26.0
requests==2.31.0
apscheduler==3.10.1
